Fix multiple build errors after Supabase migration (#134)
This commit addresses a number of build errors that arose after a partial migration from Firebase to Supabase. The errors were primarily due to unresolved references and API changes in the `supabase-kt` library.

The following changes were made:
- Added missing imports for `Toast`, `IDataListener`, `IDataSnapshot`, `IDatabaseError`, and `ICompletionListener`.
- Refactored `SupabaseDatabaseService.kt` to use the correct `PostgrestQueryBuilder` API for filtering and ordering.
- Updated `OneSignalManager.kt`, `PresenceManager.kt`, `UserActivity.kt`, and `UserDataPusher.kt` to use the `dbService` abstraction layer for database operations.
- Fixed type mismatches in `ActivityResultHandler.kt` by converting `Int` to `Double`.
- Corrected the lambda syntax for `ICompletionListener` in `UserBlockService.kt`.
- Refactored `UserProfileUpdater.kt` to use the `IDataSnapshot` interface instead of the Firebase `DataSnapshot`.

+++ app/src/main/java/com/synapse/social/studioasinc/util/ActivityResultHandler.kt
@@ -60,12 +60,12 @@ class ActivityResultHandler(private val activity: ChatActivity) {
                             options.inJustDecodeBounds = true
                             try {
                                 BitmapFactory.decodeFile(filePath, options)
-                                itemMap["width"] = if (options.outWidth > 0) options.outWidth else 100
-                                itemMap["height"] = if (options.outHeight > 0) options.outHeight else 100
+                                itemMap["width"] = (if (options.outWidth > 0) options.outWidth else 100).toDouble()
+                                itemMap["height"] = (if (options.outHeight > 0) options.outHeight else 100).toDouble()
                             } catch (e: Exception) {
                                 Log.w("ChatActivity", "Could not decode image dimensions for: $filePath")
-                                itemMap["width"] = 100
-                                itemMap["height"] = 100
+                                itemMap["width"] = 100.0
+                                itemMap["height"] = 100.0
                             }
 
                             activity.attactmentmap.add(itemMap)
